Add support for basic system metrics for Windows.
