+++ pkg/util/helpers.go
@@ -19,12 +19,8 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/cobaugh/osrelease"
-
 	"k8s.io/node-problem-detector/pkg/types"
 )
-
-var osReleasePath = "/etc/os-release"
 
 // GenerateConditionChangeEvent generates an event for condition change.
 func GenerateConditionChangeEvent(t string, status types.ConditionStatus, reason string, timestamp time.Time) types.Event {
@@ -64,39 +60,4 @@ func GetStartTime(now time.Time, uptimeDuration time.Duration, lookbackStr strin
 	}
 
 	return startTime, nil
-}
-
-// GetOSVersion retrieves the version of the current operating system.
-// For example: "cos 77-12293.0.0", "ubuntu 16.04.6 LTS (Xenial Xerus)".
-func GetOSVersion() (string, error) {
-	osReleaseMap, err := osrelease.ReadFile(osReleasePath)
-	if err != nil {
-		return "", err
-	}
-	switch osReleaseMap["ID"] {
-	case "cos":
-		return getCOSVersion(osReleaseMap), nil
-	case "debian":
-		return getDebianVersion(osReleaseMap), nil
-	case "ubuntu":
-		return getDebianVersion(osReleaseMap), nil
-	case "centos":
-		return getDebianVersion(osReleaseMap), nil
-	case "rhel":
-		return getDebianVersion(osReleaseMap), nil
-	default:
-		return "", fmt.Errorf("Unsupported ID in /etc/os-release: %q", osReleaseMap["ID"])
-	}
-}
-
-func getCOSVersion(osReleaseMap map[string]string) string {
-	// /etc/os-release syntax for COS is defined here:
-	// https://chromium.git.corp.google.com/chromiumos/docs/+/8edec95a297edfd8f1290f0f03a8aa35795b516b/os_config.md
-	return fmt.Sprintf("%s %s-%s", osReleaseMap["ID"], osReleaseMap["VERSION"], osReleaseMap["BUILD_ID"])
-}
-
-func getDebianVersion(osReleaseMap map[string]string) string {
-	// /etc/os-release syntax for Debian is defined here:
-	// https://manpages.debian.org/testing/systemd/os-release.5.en.html
-	return fmt.Sprintf("%s %s", osReleaseMap["ID"], osReleaseMap["VERSION"])
 }
